==> Part5_combining_dataframes-NoOutput.ipynb <==
# Import all dataframes for merging

import os
os.listdir()

import pandas as pd
import numpy as np

#Part1_basic_patient_info

basic = pd.read_csv('Part1_basic_patient_info.csv')
basic = basic.drop("Unnamed: 0",1)
basic

#Part2_SMI

smi = pd.read_csv('Part2_SMI_data_SMIs_Grouped.csv')
smi = smi.drop("Unnamed: 0",1)
smi

#Part3_BMI

bmi = pd.read_csv('Part3_bmi_summary_data.csv')
bmi = bmi[['PatientId', 'Baseline_BMI', 'Most_Recent_BMI', 'BMI_change_PerYear', 'Mean_BMI',]]
bmi

#Part4_Appointment History

appointments = pd.read_csv('Part4_Appointment_History.csv')
appointments = appointments[['PatientId', 'Appointments per year']]
#appointments = appointments.drop_duplicates()
appointments

#Creating master dataset

master = basic.merge(appointments,on='PatientId', how = 'left').merge(bmi,on='PatientId', how = 'left').merge(smi,on='PatientId', how = 'left')
master

from pandas import ExcelWriter

master.to_csv('master_data.csv', sep=',')
